Kernel: implement a rudimentary heap.
Implements a rudimentary slab allocator, which provides a GlobalAlloc implementation allowing the kernel
to use the collections from the alloc crate. This still needs a ton of work, will likely need to implement
a first fit allocator for cases when the requested allocation is larger than the slab allocator can
provide.